+++ src/panel/IconWidget.cc
@@ -1,6 +1,6 @@
 //
 // IconWidget.cc for pekwm
-// Copyright (C) 2022 Claes Nästén <pekdon@gmail.com>
+// Copyright (C) 2022-2023 Claes Nästén <pekdon@gmail.com>
 //
 // This program is licensed under the GNU GPL.
 // See the LICENSE file for more information.
@@ -175,4 +175,3 @@ IconWidget::parseIcon(const CfgParser::Entry* section)
 	// inital load of image to get size request right
 	load();
 }
-