+++ src/panel/WidgetFactory.hh
@@ -1,6 +1,6 @@
 //
 // WidgetFactory.hh for pekwm
-// Copyright (C) 2022 Claes Nästén <pekdon@gmail.com>
+// Copyright (C) 2022-2023 Claes Nästén <pekdon@gmail.com>
 //
 // This program is licensed under the GNU GPL.
 // See the LICENSE file for more information.
@@ -20,9 +20,12 @@
  */
 class WidgetFactory {
 public:
-	WidgetFactory(const PanelTheme& theme,
+	WidgetFactory(PWinObj* parent, Observer* observer,
+		      const PanelTheme& theme,
 		      VarData& var_data, WmState& wm_state)
-		: _theme(theme),
+		: _parent(parent),
+		  _observer(observer),
+		  _theme(theme),
 		  _var_data(var_data),
 		  _wm_state(wm_state)
 	{
@@ -31,6 +34,8 @@ public:
 	PanelWidget* construct(const WidgetConfig& cfg);
 
 private:
+	PWinObj* _parent;
+	Observer* _observer;
 	const PanelTheme& _theme;
 	VarData& _var_data;
 	WmState& _wm_state;